+++ src/Levels/Level.java
@@ -61,9 +61,9 @@ public class Level {
 
     private void loadEntities(int greenValue, int x, int y) {
         switch (greenValue) {
-            case SLIME -> slimes.add(new Slime(x * Game.TILE_SIZE, y * Game.TILE_SIZE));
-            case MUSHROOM -> mushrooms.add(new Mushroom(x * Game.TILE_SIZE, y * Game.TILE_SIZE));
-            case BUSH -> bushes.add(new Bush(x * Game.TILE_SIZE, y * Game.TILE_SIZE));
+            case SLIME ->  slimes.add(new Slime (x * (float )Game.TILE_SIZE, y * (float) Game.TILE_SIZE));
+            case MUSHROOM -> mushrooms.add(new Mushroom(x * (float) Game.TILE_SIZE, y * (float) Game.TILE_SIZE));
+            case BUSH -> bushes.add(new Bush(x * (float) Game.TILE_SIZE, y * (float) Game.TILE_SIZE));
             case 100 -> playerSpawn = new Point(x * Game.TILE_SIZE, y * Game.TILE_SIZE);
         }
     }
